requirements.txt
django
zencore-django
click
